Test shader extension restoration on context lost
This adds tests against OES_standard_derivatives and EXT_shder_lod after context lost/restore occurs.

Figma detected problems in WebKit context restoration with regard to shader extensions.
https://bugs.webkit.org/show_bug.cgi?id=177270

This existing tests are for OES_vertex_array_object which isn't supported by IE, and OES_texture_float which isn't supported on older mobile devices.  Both of these fail to restore in Safari 11.0, so these shader extensions fail in addition to the original extension tests. Any WebGL implementation should not be considered a 1.0 implementation if a caller can't properly recover from context lost in this way.  Shaders that did compile the first time, fail after context lost.  This code follows the convention that the extensions must first be present in order to restore them.  It's hard to say if context lost events from prior conformance tests might then lead to this test not receiving the extensions in the first place, but I ran this test in isolation.